refactor(ui): split useTypingTest into matrix, metrics, and memory modules
